Filesize can be more than 2GB, woops
git-svn-id: svn://svn.madcowdisease.org/mcd/blamehangle/trunk@3689 35f662cf-58f2-0310-8820-d172d141b225

+++ schemas/TorrentScraper.sql
@@ -2,5 +2,5 @@ CREATE TABLE torrents (
 	scrape_time integer NOT NULL,
 	url character varying(255) NOT NULL PRIMARY KEY,
 	filename character varying(255),
-	filesize integer
+	filesize bigint
 );
